requirements.txt
django~=4.0.3
psycopg2-binary==2.8.5
gunicorn==20.1.0
pytz==2022.1
djangorestframework==3.12.4